Remove a bunch more warning errors for unused variables.  Only two source remain:  in OptimizeProblem.cpp where all input arguments are unused at this point  in dotref.cpp where the timer for allreduce is unused in serial compilation mode.
M    src/SetupHalo.cpp
M    src/SymTest.cpp
M    src/YAML_Element.cpp
M    src/WriteProblem.cpp
M    src/GenerateProblem.cpp
M    src/SymTest.hpp
M    src/YAML_Element.hpp
M    src/OptimizeProblem.cpp
M    src/ExchangeHalo.cpp
M    src/CG.cpp
M    src/Geometry.hpp
M    src/GenerateGeometry.cpp
M    src/CGref.cpp
M    src/SparseMatrix.hpp
M    src/CGtest.cpp
M    setup/Make.Linux_Serial
M    testing/main.cpp


git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@110 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/GenerateGeometry.cpp
@@ -97,7 +97,7 @@ Counter_max_cur_sub(Counter *this_, Counter *that, Counter *res) {
 
 static void
 primefactor_i(int x, int *factors) {
-  int i, d, sq=(int)(sqrt(x))+1L;
+  int i, d, sq=(int)(sqrt((double)x))+1L;
   div_t r;
 
   /* remove 2 as a factor with shifts */
